GH-1365 correctly check for null

+++ libraries/chain/controller.cpp
@@ -2147,7 +2147,7 @@ struct controller_impl {
       if (conf.truncate_at_block > 0 && chain_head.is_valid()) {
          if (chain_head.block_num() == conf.truncate_at_block && fork_db_has_root()) {
             fork_db_.apply<void>([&](auto& fork_db) {
-               if (auto head = fork_db.head(); head->block_num() > conf.truncate_at_block) {
+               if (auto head = fork_db.head(); head && head->block_num() > conf.truncate_at_block) {
                   ilog("Removing blocks past truncate-at-block ${t} from fork database with head at ${h}",
                         ("t", conf.truncate_at_block)("h", head->block_num()));
                   fork_db.remove(conf.truncate_at_block + 1);
